refactor: add SessionCallbackAdaptor of the TurboSessionCallback
